fix: merged federation and metadata policy implementation
