--- pyeudiw/metadata/exceptions.py
@@ -1,6 +1,0 @@
-__author__ = "Roland Hedberg"
-__license__ = "Apache 2.0"
-__version__ = ""
-
-class PolicyError(Exception):
-    pass
